% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{check_barcodes}
\alias{check_barcodes}
\title{Given a vector of sequences, check barcodes by aligning}
\usage{
check_barcodes(sequences, set = c("a", "alpha", "het", "hom"))
}
\arguments{
\item{sequences}{Vector of barcode sequence results (i.e. vector of strings of 20 nucleotides).}

\item{set}{One or more of \code{c('a', 'alpha', 'het', 'hom')}. Which collection of barcodes to check.}
}
\description{
Given a vector of sequences, check barcodes by aligning
}
