% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_barcodes}
\alias{download_barcodes}
\title{Download barcodes}
\usage{
download_barcodes(to = ".", set = c("a", "alpha", "het", "hom"))
}
\arguments{
\item{to}{directory path. Where to download barcodes.}

\item{set}{One or more of \code{c('a', 'alpha', 'het', 'hom')}. Which collection of barcodes to download.}
}
\value{
Returns a dataframe with five columns.

\itemize{
  \item{strain_id}{An ID for this strain. Begins with the characters "rec" followed by the record number.}
  \item{orf}{The original systematic ORF designation for this strain. Note that a few ORF IDs have been removed or changed.}
  \item{uptag}{The 20 NT uptag sequence for the corresponding strain.}
  \item{dntag}{The 20 NT dntag sequence for the corresponding strain.}
  \item{type}{Type of strain. One of "a", "alpha", "het", or "hom" for MATa, MATalpha, Heterozygous diploid and Homozygous diploid.}
}
}
\description{
Download barcodes from the \href{http://www-sequence.stanford.edu/group/yeast_deletion_project/deletions3.html}{yeast gene deletion consortium}
}
