% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-barcodes.R
\name{src_barcodes}
\alias{src_barcodes}
\title{Connect to the barcodes database}
\usage{
src_barcodes()
}
\description{
Connect to the barcodes SQLite database.
}
\examples{

# Connect to the database
db <- src_barcodes()
db

}
